deps: Add set -x for installation commands
Adds set -x for all installation commands in install_requirements.sh so
that users can see what's actually being installed and can help debug
when they run into any issues.

NOTE: This doesn't turn it on for all of the script (since that'd be
overbearing) but rather turns it on specifically when installing things.

This also adds "set -eou pipefail" to the top of the script which is a
recommended paradigm for all bash scripts since it adds things like
error on exit, error on undefined variable, and propagation of errors on
pipe failure.

https://explainshell.com/explain?cmd=set+-euo+pipefail

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>