Download fix (#1366)
* fix: allow multiple weight mapping files for mistral

Downloading a Mistral model fails because it includes multiple weight
mapping files. The regression was introduced in commit
`766bee9f4a1fcb187fae543a525495d3ff482097`. I'm unclear on the original
intent, but perhaps the exception was meant to apply only to Granite
models. This isn’t an ideal fix, but it does enable Mistral to be
downloaded and used for chat.

Signed-off-by: Sébastien Han <seb@redhat.com>

* fix(download): Fix safetensors/bin/pth download logic

The previous logic didn't handle .bin files, so if a model (like mistral)
has both .bin and .safetensors, it would download both.

Branch: download-fix

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

* fix(convert hf): Better logic to handle multiple weight mapping files

This will not actually be needed for mistral with the fix in download to
handle .bin files, but it may be needed for other models, so it's worth
having.

Branch: download-fix

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>

---------

Signed-off-by: Sébastien Han <seb@redhat.com>
Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>
Co-authored-by: Sébastien Han <seb@redhat.com>