Force run-readme-pr-macos.yml to use CPU instead of incorrectly loading to MPS (#1417)
* bandaid for run-readme-pr-macos.yml incorrectly loading to MPS

as per #1416 torchchat on hosts without MPS (which is all github hosts which use kvm to virtualize MacOS, but not MPS) should choose CPU as "fast" device.  The logic is present (see discussion in #1416 ), but either not fully functional (that would be the easier one to fix, just print the result of get_device_str and fix the code!) or specifically ignored on load in torch/serialization.py (If this is the case, we're effectively looking at a core PyTorch issue....)

In the meantime, this bandaid just forces the use of CPU on MacOS tests, to make MacOS tests run on CPU -- labeit hsortcircuiting test/execution of the "fast" device logic.  Not ideal, but some testing beats no testing.

* Update run-readme-pr-macos.yml

Add informational message to MacOS CPU tests

* Update build_native.sh

Update to C++11 ABI for AOTI, similar to ET

---------

Co-authored-by: Jack-Khuu <jack.khuu.7@gmail.com>